Refactoring JUnit tests (#82)
* Add JUnit dependency in build.gradle

* Refactoring JUnit tests and Java testcases structure

* Removed some comments

* Removed test

* Always serialize results in tests

* Fixed allInstances method of Java's IntType

* Fixed test02

* Number literals always store an integer value

* Fixing typing test.

* Fixed prefix and postfix addition and subtraction

* Class static type is modeled as a reference type

* Fixed return test

* Fixed prefix-postfix test

* Fixed parenthesized-expr test

* Fixed loop tests

* Removed FIXME comments for fixed tests

* Refactoring of addition

* Fixed assignment

* Simplified Java assignment (relying now on runtime types of the
right-hand expression)

* Simplified and fixed Java add

* Refactoring tests

* Fixing test01

* Minor changes

* Removed files

* Removed some tests

* Fixed AssignmentOperator test

* Fixed AssertTest test

* Fixed DefaultConstructor test

* Fixed SuperConstructor1 test

* Changed assert test

* Refactoring test file names

* Removed test file

---------

+++ jlisa/src/main/java/it/unive/jlisa/frontend/visitors/ExpressionVisitor.java
@@ -544,11 +544,11 @@ public class ExpressionVisitor extends JavaASTVisitor {
         try {
             long value = Long.decode(token); // handles 0x, 0b, octal, decimal
             if (value >= Byte.MIN_VALUE && value <= Byte.MAX_VALUE) {
-                expression = new ByteLiteral(this.cfg, getSourceCodeLocation(node), (byte)value);
+                expression = new ByteLiteral(this.cfg, getSourceCodeLocation(node), (int) value);
             } else if (value >= Short.MIN_VALUE && value <= Short.MAX_VALUE) {
-                expression = new ShortLiteral(this.cfg, getSourceCodeLocation(node),(short)value);
+                expression = new ShortLiteral(this.cfg, getSourceCodeLocation(node),(int) value);
             } else if (value >= Integer.MIN_VALUE && value <= Integer.MAX_VALUE) {
-                expression = new IntLiteral(this.cfg, getSourceCodeLocation(node),(int)value);
+                expression = new IntLiteral(this.cfg, getSourceCodeLocation(node), (int) value);
             } else {
                 expression = new LongLiteral(this.cfg, getSourceCodeLocation(node), value);
             }
